requirements.txt
pandas
tqdm
matplotlib
better-profanity
python-dotenv
requests
